Внесение изменений в задание 45/46

+++ app/src/main/java/com/example/android_intern/presentation/model/ForecastUI.kt
@@ -2,6 +2,6 @@ package com.example.android_intern.presentation.model
 
 data class ForecastUI(
     val dtTxt: String,
-    val temp: Double,
+    val temp: Int,
     val icon: String
 )
